+++ src/main/java/org/folio/search/utils/KafkaConstants.java
@@ -9,6 +9,7 @@ public final class KafkaConstants {
   public static final String CLASSIFICATION_TYPE_LISTENER_ID = "mod-search-classification-type-listener";
   public static final String LOCATION_LISTENER_ID = "mod-search-location-listener";
   public static final String BIBFRAME_LISTENER_ID = "mod-search-bibframe-listener";
+  public static final String CAMPUS_LISTENER_ID = "mod-search-campus-listener";
 
   private KafkaConstants() {}
 }